实现回复功能-异常处理

+++ src/main/java/file/majing/community/controller/PublishController.java
@@ -18,7 +18,7 @@ import javax.servlet.http.HttpServletRequest;
 @Controller public class PublishController {
 	@Autowired private QuestionService questionService;
 	@GetMapping("/publish/{id}")
-	public String edit(@PathVariable(name="id")Integer id,Model model){
+	public String edit(@PathVariable(name="id")Long id,Model model){
 		QuestionDTO question=questionService.getById(id);
 		model.addAttribute("title", question.getTitle());
 		model.addAttribute("description", question.getDescription());
@@ -35,7 +35,7 @@ import javax.servlet.http.HttpServletRequest;
 			@RequestParam(value = "title", required = false) String title,
 			@RequestParam(value = "description", required = false) String description,
 			@RequestParam(value = "tag", required = false) String tag,
-			@RequestParam(value = "id",required =false)Integer id,
+			@RequestParam(value = "id",required =false)Long id,
 			HttpServletRequest request, Model model) {
 		model.addAttribute("title", title);
 		model.addAttribute("description", description);
